requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.3.0
appdirs==1.4.4
APScheduler==3.10.4
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cssselect==1.2.0
dateparser==1.2.0
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
fake-useragent==1.4.0
fastapi==0.110.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.4
httptools==0.6.1
httpx==0.27.0
idna==3.6
importlib-metadata==7.0.1
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.1.0
magic-filter==1.0.12
MarkupSafe==2.1.5
multidict==6.0.5
openai==1.12.0
orjson==3.9.15
outcome==1.3.0.post0
parse==1.20.1
pillow==10.2.0
pydantic==2.5.3
pydantic-extra-types==2.6.0
pydantic-settings==2.2.1
pydantic_core==2.14.6
pyee==11.1.0
pymongo==4.6.2
pyppeteer==2.0.0
pyquery==2.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
requests-html==0.10.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.36.3
tqdm==4.66.2
trio==0.24.0
trio-websocket==0.11.1
typing_extensions==4.9.0
tzlocal==5.2
ujson==5.9.0
urllib3==1.26.18
uvicorn==0.29.0
uvloop==0.19.0
w3lib==2.1.2
watchfiles==0.21.0
websockets==10.4
wsproto==1.2.0
yarl==1.9.4
zipp==3.17.0
